chore(structure): move `web` from `apps` to project root and rename it to `site` (#44)
* chore: move apps/web to site

this commit moves the entire `apps/web` directory to `site`.
this change was necessary to restructure the project layout,
likely as part of a larger architectural change or a move
towards a more unified codebase. this refactoring doesn't
introduce any breaking changes.

* chore: update pnpm version to 10.11.1

updates pnpm version to the latest patch release to ensure
compatibility and take advantage of bug fixes.

this change is necessary to keep the project up-to-date with
the latest tooling. no breaking changes are introduced.