release: `v0.0.1` (#45)
* refactor(style): refactor styling from `vanilla-extract` to `tailwindcss` (#43)

* refactor: remove constants file and inline default icon size

* feat: add base styles, animation, and tailwind integration

* refactor: move component packages to app directory

* chore: update eslint config and dependencies

* chore: integrate tailwind css and remove vanilla-extract

* chore: remove vanilla-extract and use tailwind css

* feat: add backdrop blur to the top navigation overlay

* refactor: move components and inline styles

* chore: remove unused css files and update styles

* refactor: inline styles in project page and crafts page

* refactor: move project page styles to tailwind

* refactor: remove css modules and use inline styles in resume page

* refactor: apply tailwind css

This commit refactors the styling of several components
to use Tailwind CSS classes instead of vanilla-extract.
This change was necessary to move toward a more
consistent and maintainable styling approach.

The following components were updated:
- BackButton
- RefreshButton
- Signature
- Timeline

This change removes the corresponding `styles.css.ts`
files for these components.
There are no breaking changes.

* chore: remove vanilla extract and migrate to tailwind

* chore: remove vanilla extract and migrate to tailwind

* chore: remove husky and lint-staged

remove husky and lint-staged from the project.

this commit removes the husky and lint-staged packages. they are no
longer needed as the project's formatting and linting are now handled
by other tools. this simplifies the project's dependencies and
build process. there

* style: refine article list item layout

This commit refines the layout of the article list items to improve responsiveness and visual consistency across different screen sizes.

- Removed `items-center` from the Link component's classname, as it was causing unwanted vertical alignment on larger screens.
- Added `self-start` and `mobile:self-center` to the time element's classname to align the time to the top on smaller screens and center it on larger screens.

These changes ensure that the article list items adapt gracefully to various screen sizes, providing a better user experience.

* chore: remove article.css.ts, update styles and content

- removed `article.css.ts` as styles are now defined in `utilities.css`
- moved article styles from `article.css.ts` to `utilities.css`
- updated `globals.css` to import `theme.css` instead of `theme2.css`
- updated `index.mdx` post
- update components: mdx component

* feat: apply tailwind css and update image component

This commit refactors the `utilities.css` file to use Tailwind CSS classes instead of custom CSS. This improves maintainability and consistency.
Additionally, the `Image` component in `mdx-component.tsx` is updated to be simpler.

The changes include:

- Replacing custom CSS rules with Tailwind CSS classes in `utilities.css`.
- Removing the `span` wrapper and object-contain style from the `Image` component in `mdx-component.tsx`, and adding the sizes prop.
- Updating image links in `opensource-and-coderplace/index.mdx` to use a direct link.

These changes enhance the styling consistency and simplify the image handling.

* chore(structure): move `web` from `apps` to project root and rename it to `site` (#44)

* chore: move apps/web to site

this commit moves the entire `apps/web` directory to `site`.
this change was necessary to restructure the project layout,
likely as part of a larger architectural change or a move
towards a more unified codebase. this refactoring doesn't
introduce any breaking changes.

* chore: update pnpm version to 10.11.1

updates pnpm version to the latest patch release to ensure
compatibility and take advantage of bug fixes.

this change is necessary to keep the project up-to-date with
the latest tooling. no breaking changes are introduced.

* style: remove inline padding from body

This change removes the `padding-inline` from the body style.

The `padding-inline` was causing layout issues on smaller screens. By removing it, the layout is more responsive and adapts better to different screen sizes.

This change does not introduce any breaking changes.

* feat: allow optional article descriptions

This commit modifies the article display to support optional descriptions.

- In the mdx content files, the description field is now optional.
- The articles list and individual article pages have been updated to conditionally render the description only if it exists.
- This allows for articles that might not require a detailed summary.

* style: improve code highlight styling

Updates code highlight styles for better appearance and consistency.

- Changes border color of code blocks to use `--color-card-border` variable
- Updates light theme code color to `--color-gray950`
- Updates dark theme code background color to `--color-gray950`
- Updates code font size to `text-sm` and leading to `leading-normal`
- Updates line number color to `--color-text-secondary`
- Use tailwind's apply directive to set the hidden style to code blocks

* feat: enhance image loading with blur placeholders

This commit introduces blur placeholders for images in articles,
improving the perceived performance and user experience during
page load.

The changes involve:

- Adding `sharp` dependency for image processing.
- Implementing image blur generation using `sharp`.
- Modifying the `Article` content layer to include a `blurMap`
  field that stores blur data URLs for each image.
- Updating the `MdxComponent` to display blur placeholders
  while images are loading.
- Refactoring image syntax within mdx files to utilize markdown format

This change enhances the loading experience and provides a visual
cue that content is on the way. There are no breaking changes.

* style: remove redundant image and video styling in mdx articles

This commit removes some redundant styling rules from the `utilities.css` file related to images and videos within MDX articles.

The `p > img, video` styling was overly generic and unnecessary as the default styling from the base stylesheet was sufficient.

The `.image-wrapper p` styling was also redundant as it was overriding default p tag styling without a clear benefit.

Removing these styles simplifies the CSS and reduces potential conflicts. There are no breaking changes as the default styling remains appropriate.

* chore: update dependencies to latest versions

* fix: update theme handling to use resolvedTheme in Giscus and Header components

* fix: update list item text color in MDX articles

* feat: add Craft document type and implement new crafts in content layer

* fix: correct variable name for Pretendard font in layout component

* fix: update publisher and URLs to use constants for BASE_URL and PATH

* fix: update links and improve icon handling in page component

* fix: update text color for project team in page component

* fix: improve theme handling in MapIframe component and update iframe URL

* feat: add RSS feed generation with dynamic content and site language support

* fix: add type annotations for Article and Craft in route.ts

* feat: enhance image handling in MdxComponent with aspect ratio support

* feat: enhance image blur handling with aspect ratio support in MdxComponent

* fix: update creation dates in index.mdx and enhance hr styling in page.tsx